Rename ScopeBar to follow convention
